requirements.txt
numpy>=1.16.0
trimesh>=2.38.0
tqdm>=4.50.0
scipy>=1.3.0
pandas>=0.24.2
networkx>=2.4
scikit-learn>=0.23
python-igraph>=0.8
